so it seems milestones aren't supported in the REST API
